# static/openapi.yaml
openapi: 3.0.3
info:
  title: GPT API
  description: |-
    RAG LLM API capable to conversate with context memory, it is integrated with custom company's knowledge.
    It uses a Conversational Summary Memory technique stored in a VectorDB.
  termsOfService: https://www.bosch-softwaretechnologies.com/en/our-company/about-us/
  contact:
    email: tjo1ga@bosch.com
  license:
    name: License BGSW-MX
    url: https://www.bosch-softwaretechnologies.com/en/our-company/about-us/
  version: 1.0.0
...
servers:
  - url: https://service-gpt-api.bosch-warehouse.com/api/v1
  - url: http://localhost:8000/api/v1
...
paths:
...
  /email/send-email:
    post:
      tags:
        - email
      summary: Return results of email api
      description: Return success or failure of email trigger event
      parameters:
        - name: user
          in: query
          required: true
          description: User to find.
          schema:
            type: string
        - name: conversationID
          in: query
          required: false
          description: Conversation id to find.
          schema:
            type: string
      operationId: send_email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequestModel'
            example:
              email: "khajanchandra.pandey@mx.bosch.com"
              summary: "Test subject"
      responses:
        '201':
          description: Returns response if success.
          content:
            application/json:
              schema:
                type: object
                example:
                  videoId: "Email sent successfully"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                example:
                  error: "Failed to send email."
components:
  schemas:
...
    EmailRequestModel:
      type: object
      properties:
        user:
          type: string
          description: User to find.
        email:
          type: string
          description: Email to be sent
          example: "khajanchandra.pandey@mx.bosch.com"
        summary:
          type: string
          description: Subject of email
          example: "This is test for customer service"
      required:
        - user
        - email
        - summary
...